Backend Flask/requirements.txt
Flask
Flask-CORS
google-generativeai
opencv-python
Pillow
pdf2image
datasets
huggingface_hub
python-dotenv
paddleocr
pydantic
typing-extensions
requests